feat(portal): add role-based auth flows and split portal into domain pages
- Add /register (3-step wizard: account → place → success) for user self-signup
- Add /invite/:token (InviteAcceptPage) for viewer onboarding via shareable link
- Add ViewersPage: list viewers, generate 7-day invite link, deactivate viewers
- Add PlanPage: active subscription display + upgrade request flow
- Split 1,253-line portal-app.tsx into pages/auth, pages/cameras, pages/events,
  pages/realtime, pages/account, pages/viewers, pages/plan and components/
- ProtectedLayout now role-aware: monitor sees Viewers + Plan nav, client_user sees
  only cameras/events/realtime/households/profile, tenant_admin sees Viewers
- Shared lib/client.ts extracts usePortalClient, formatApiError, constants
